Merge branch 'main' of https://github.com/KwiatQIM/Flask-App
Removed the entire results page and condensed all of the results onto the single FrontPage.html file. Edited base.html to allow for returning to the home page. Added QuantumTomography to the requirements.txt file. Added single qubit support, but the DA visibility calculations for single qubits is broken at the moment.